Fix (s32) pointer truncation in memory.c, code_5DBC0.c, particle.c, code_C89C0.c

+++ src/core2/code_5DBC0.c
@@ -549,7 +549,7 @@ void func_802E6820(s32 arg0) {
                     chunk_count = sprite_getFramePtr(prev_sprite_ptr, 0U)->chunkCnt;
                     D_8037E900->unk4[var_s5].font_bin = (BKSprite *)defrag_asset(D_8037E900->unk4[var_s5].font_bin);
                     for(i_chunk = 0; i_chunk < chunk_count; i_chunk++){
-                        D_8037E900->unk4[var_s5].letter_texture[i_chunk] = ((u32)(((s32)D_8037E900->unk4[var_s5].letter_texture[i_chunk] - (s32)prev_sprite_ptr)) + (u32)D_8037E900->unk4[var_s5].font_bin);
+                        D_8037E900->unk4[var_s5].letter_texture[i_chunk] = ((u32)(((uintptr_t)D_8037E900->unk4[var_s5].letter_texture[i_chunk] - (uintptr_t)prev_sprite_ptr)) + (u32)D_8037E900->unk4[var_s5].font_bin);
                     }
                 }
             }
